+++ pdm_api/views/default.py
@@ -20,39 +20,6 @@ def my_view(request):
         return Response(db_err_msg, content_type='text/plain', status=500)
     return {'one': one, 'project': 'pdm_api'}
 
-@view_config(route_name='slack', renderer='../templates/slack.jinja2')
-def api(request):
-    if request.method == 'GET':
-        return(HTTPForbidden())
-
-    required_keys = [
-        'token',
-        'team_id',
-        'team_domain',
-        'service_id',
-        'channel_id',
-        'cannel_name',
-        'user_id',
-        'user_name',
-        'text'
-    ]
-
-    remote_addr = str(request.remote_addr)
-    check(request, remote_addr)
-
-    for record in required_keys:
-        if record not in request.params:
-            return(HTTPForbidden())
-
-    header_values = {}
-
-    for record  in request.POST.items():
-        header_values[str(record[0])] = str(record[1])
-
-    add(request, remote_addr, str(header_values))
-
-    return(header_values)
-
 
 db_err_msg = """\
 Pyramid is having a problem using your SQL database.  The problem
